Refactor to modular routes and utilities
