+++ config/adminCredentials.js
@@ -1,0 +1,4 @@
+module.exports = {
+  username: 'admin',
+  passwordHash: '$2b$10$R4tMQGVYYReQayD82yx.6.E/4bE.0Ue.vmmWT6t1ggXrJFA3wUCqu'
+};
